Version 1.1 z błedami w tablicy

+++ src/Point.java
@@ -1,24 +1,24 @@
 public class Point {
-    private int coordinateX = 0;
-    private int CoordinateY = 0;
+    private double coordinateX = 0;
+    private double CoordinateY = 0;
 
     public Point(){
 
     }
 
-    public int getCoordinateX() {
+    public double getCoordinateX() {
         return coordinateX;
     }
 
-    public void setCoordinateX(int coordinateX) {
+    public void setCoordinateX(double coordinateX) {
         this.coordinateX = coordinateX;
     }
 
-    public int getCoordinateY() {
+    public double getCoordinateY() {
         return CoordinateY;
     }
 
-    public void setCoordinateY(int coordinateY) {
+    public void setCoordinateY(double coordinateY) {
         CoordinateY = coordinateY;
     }
 }
